Playing with CSS
Took some css from the intro sveltekit app. Also played around with the footer, specifically to get them all to fit and stay in the same line. Turns out the problem was flex-direction: column in the footer. Got rid of that and they all lined up.

Not going to keep it green. At least not this green, probably a similar green. I like green.